pybot/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
flake8==7.0.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
pyTelegramBotAPI==4.15.4
pytest==8.1.1
pytest-isort==4.0.0
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.0
